% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_between_vectors.R
\name{cos_between_vectors}
\alias{cos_between_vectors}
\title{Find cosine of angle between vectors}
\usage{
cos_between_vectors(vector1, vector2, verbose = FALSE)
}
\arguments{
\item{vector1}{Numeric vector}

\item{vector2}{Numeric vector}

\item{verbose}{boolean}
}
\value{
Double representing the cosine of the angle between the given vectors.
}
\description{
Find cosine of angle between vectors
}
\examples{
cos_between_vectors(c(0, 1), c(1, 0))
}
